+++ game/source/ai/ai_actions.hpp
@@ -5,19 +5,19 @@
 struct c_ai_point3d
 {
 	real_point3d m_point;
-	short m_reference_frame;
-	short m_bsp_index;
+	int16 m_reference_frame;
+	int16 m_bsp_index;
 };
 static_assert(sizeof(c_ai_point3d) == 0x10);
 
 struct c_ai_direction
 {
-	long m_type;
+	int32 m_type;
 
 	union
 	{
-		long m_pref_index;
-		long m_object_index;
+		int32 m_pref_index;
+		int32 m_object_index;
 		real_point3d m_point;
 		real_vector3d m_vector;
 		c_ai_point3d m_ai_point;
@@ -27,19 +27,19 @@ static_assert(sizeof(c_ai_direction) == 0x14);
 
 struct c_ai_action
 {
-	long m_ai_action;
+	int32 m_ai_action;
 	bool m_point_valid;
 	uint8 m_pad0;
-	short m_ticks;
+	int16 m_ticks;
 
 	union
 	{
-		long m_data;
-		long m_animation_name;
-		long m_posture_name;
+		int32 m_data;
+		int32 m_animation_name;
+		int32 m_posture_name;
 		struct
 		{
-			short m_corner_type;
+			int16 m_corner_type;
 			bool m_open;
 		} m_corner;
 	};
