+++ game/source/ai/ai_script.hpp
@@ -10,5 +10,5 @@ extern void __cdecl ai_script_initialize();
 extern void __cdecl ai_script_initialize_for_new_map();
 extern void __cdecl ai_scripting_erase_all();
 
-extern bool ai_index_from_string(struct scenario* scenario, char const* ai_string, long* ai_index_reference);
+extern bool ai_index_from_string(struct scenario* scenario, char const* ai_string, int32* ai_index_reference);
